custom_bucket/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
Pillow==10.4.0
pydantic==2.9.2
python-magic==0.4.27
aiofiles==24.1.0
